requirements.txt
fastapi==0.115.5
pydantic==2.9.2
uvicorn==0.32.0
python-dotenv==1.0.1
sqlmodel==0.0.22
pyjwt==2.9.0
passlib==1.7.4
bcrypt==4.2.0